% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto-dashboard.R
\name{gh_dashboard}
\alias{gh_dashboard}
\title{Start quarto dashboard with results of main \link{githist} function.}
\usage{
gh_dashboard(results)
}
\arguments{
\item{results}{Results of main \link{githist} function applied to one
package.}
}
\value{
Nothing; quarto dashboard will automatically open on function call.
}
\description{
Start quarto dashboard with results of main \link{githist} function.
}
