% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githist-package.R
\docType{package}
\name{githist-package}
\alias{githist-package}
\title{githist: Code analyses traced along the 'git' history of a package}
\description{
Code analyses traced along the 'git' history of a package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/githist/}
  \item \url{https://github.com/ropensci-review-tools/githist}
  \item Report bugs at \url{https://github.com/ropensci-review-tools/githist/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com} (\href{https://orcid.org/0000-0003-2172-5265}{ORCID})

}
\keyword{internal}
