% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githist-fn.R
\name{githist}
\alias{githist}
\title{Apply \pkg{pkgstats} across the git history of a package}
\usage{
githist(path, n = NULL)
}
\arguments{
\item{path}{Path to local repository containing an R package.}

\item{n}{If given, only analyses the preceding 'n' commits in the git
history.}
}
\description{
Apply \pkg{pkgstats} across the git history of a package
}
