% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto-dashboard-org.R
\name{orgmetrics_dashboard}
\alias{orgmetrics_dashboard}
\title{Start quarto dashboard with results of
\code{repometrics_collate_org_data} function for collation of data across orgs.}
\usage{
orgmetrics_dashboard(data_org, fn_calls, emb_matrix, action = "preview")
}
\arguments{
\item{data_org}{Data on GitHub organization as returned from
\code{repometrics_collate_org_data} function.}

\item{fn_calls}{Data on function calls between packages of the specified
organization, as returned from the \code{rm_org_data_fn_call_network()} function.}

\item{emb_matrix}{Matrix of embeddings between packages, returned from
\code{rm_org_emb_distances()} function.}

\item{action}{One of "preview", to start and open a live preview of the
dashboard website, or "render" to render a static version without previewing
or opening.}
}
\value{
(Invisibly) Path to main "index.html" document of quarto site. Note
that the site must be served with \code{action = "preview"}, and will not work by
simply opening this "index.html" file.
}
\description{
Start quarto dashboard with results of
\code{repometrics_collate_org_data} function for collation of data across orgs.
}
