% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chaoss-external.R
\name{prop_commits_in_change_req}
\alias{prop_commits_in_change_req}
\title{The "Ratio of Code Commits linked with Change Requests" CHAOSS metric. This
is defined as, "Percentage of new code commits linked with change requests
in the last 90 days."
\url{https://chaoss.community/kb/metrics-model-collaboration-development-index/}.}
\usage{
prop_commits_in_change_req(path, end_date = Sys.Date())
}
\description{
The "Ratio of Code Commits linked with Change Requests" CHAOSS metric. This
is defined as, "Percentage of new code commits linked with change requests
in the last 90 days."
\url{https://chaoss.community/kb/metrics-model-collaboration-development-index/}.
}
