% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pkgstats.R
\name{repo_pkgstats_history}
\alias{repo_pkgstats_history}
\title{Apply \pkg{pkgstats} across the git history of a package}
\usage{
repo_pkgstats_history(path, step_days = 1L, num_cores = -1L)
}
\arguments{
\item{path}{Path to local repository containing an R package.}

\item{step_days}{Analyse package at intervals of this number of days. The
last commit for each day is chosen. For example, \code{step_days = 7L} will
return weekly statistics. Values of zero or less will analyse all commits,
including potentially multiple daily commits.}

\item{num_cores}{Number of cores to use in multi-core processing. Has no
effect on Windows operating systems, on which calculations are always
single-core only. Negative values are subtracted from number of available
cores, determined as \code{parallel::detectCores()}, so default of \code{num_cores = -1L} uses \code{detectCores() - 1L}. Positive values use precisely that number,
restricted to maximum available cores, and a value of zero will use all
available cores.}
}
\value{
\code{NULL} if \code{path} is not an R package, or if no \pkg{pkgstats}
results are able to be extracted. Otherwise, a list of three items:
\itemize{
\item desc_data Containing data from \code{DESCRIPTION} files, along with data on
numbers of functions.
\item loc Containing data on "lines-of-code" for all languages and
sub-directories within package.
\item stats Containing statistics on (mean, medium, and sum) of various
properties of each function in package.
}
}
\description{
Apply \pkg{pkgstats} across the git history of a package
}
