% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repometrics-package.R
\docType{package}
\name{repometrics-package}
\alias{repometrics}
\alias{repometrics-package}
\title{repometrics: Metrics for Your Code Repository}
\description{
Metrics for your code repository. Call one function to generate an interactive dashboard displaying the state of your code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci-review-tools/repometrics}
  \item Report bugs at \url{https://github.com/ropensci-review-tools/repometrics/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com} (\href{https://orcid.org/0000-0003-2172-5265}{ORCID})

}
\keyword{internal}
