% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-repo.R
\name{repometrics_data}
\alias{repometrics_data}
\title{Collate 'repometrics' data for a local R package.}
\usage{
repometrics_data(
  path,
  date_interval = "month",
  num_cores = -1L,
  end_date = Sys.Date(),
  nyears = 1
)
}
\arguments{
\item{path}{Path to local repository containing an R package.}

\item{date_interval}{Interval at which to analyse package statistics, with
options of "day", "week", "month", or "year". The last commit for each
period is chosen.}

\item{num_cores}{Number of cores to use in multi-core processing. Has no
effect on Windows operating systems, on which calculations are always
single-core only. Negative values are subtracted from number of available
cores, determined as \code{parallel::detectCores()}, so default of \code{num_cores = -1L} uses \code{detectCores() - 1L}. Positive values use precisely that number,
restricted to maximum available cores, and a value of zero will use all
available cores.}

\item{end_date}{Parameter used in some aspects of resultant data to limit
the end date of data collection. Defaults to \code{Sys.Date ()}.}

\item{nyears}{Parameter <= 1 determining fraction of a year over which data
up until \code{end_date} are collected.}
}
\value{
A list of five items:
\enumerate{
\item "pkgstats" containing statistics on the historical development of
package code, derived from the \pkg{pkgstats} package;
\item "rm" containing data from GitHub on the repository, including data on
contributors, issues, pull requests, and people watching and starring the
repository.
\item "contributors" as a named list of data on every individual contributor
to the repository, whether by code contributions or GitHub issues or
discussions.
\item "rm_metrics" as a list of values for all CHAOSS metrics defined in the
output of \link{rm_chaoss_metrics_list}.
\item "rm_models" as a list of values for CHAOSS models, derived from
aggregating various metrics.
}
}
\description{
This function collates all data for a local R package or repository needed
to create a dashboard with the \link{repometrics_dashboard} function. It
combines data from both the \link{repometrics_data_repo} and
\link{repometrics_data_user} functions.
}
\seealso{
Other data: 
\code{\link{repo_pkgstats_history}()},
\code{\link{repometrics_data_repo}()},
\code{\link{repometrics_data_user}()}
}
\concept{data}
