% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-repo.R
\name{repometrics_data_repo}
\alias{repometrics_data_repo}
\title{Collate 'repometrics' data for a local R package}
\usage{
repometrics_data_repo(path, step_days = 1L, num_cores = -1L)
}
\arguments{
\item{path}{Path to local repository containing an R package.}

\item{step_days}{Analyse package at intervals of this number of days. The
last commit for each day is chosen. For example, \code{step_days = 7L} will
return weekly statistics. Values of zero or less will analyse all commits,
including potentially multiple daily commits.}

\item{num_cores}{Number of cores to use in multi-core processing. Has no
effect on Windows operating systems, on which calculations are always
single-core only. Negative values are subtracted from number of available
cores, determined as \code{parallel::detectCores()}, so default of \code{num_cores = -1L} uses \code{detectCores() - 1L}. Positive values use precisely that number,
restricted to maximum available cores, and a value of zero will use all
available cores.}
}
\value{
A list with two main items:
\enumerate{
\item "pkgstats" Containing summary data from apply \code{pkgstats} routines
across the git history of the repository.
\item "cm" Containing data used to derive "CHAOSS metrics", primarily from
GitHub data.
}
}
\description{
Collate 'repometrics' data for a local R package
}
