% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm-collate-all.R
\name{rm_chaoss_metrics_list}
\alias{rm_chaoss_metrics_list}
\title{List all implemented CHAOSS metrics}
\usage{
rm_chaoss_metrics_list()
}
\value{
A character vector with the internal function names of all
implemented CHAOSS metrics.
}
\description{
This function returns a list of internal functions defined within the
'repometrics' package. These internal functions are not intended to be
called directly, rather this list is provided for information only, to
enable users to know which metrics are implemented.
}
\examples{
metrics <- rm_chaoss_metrics_list ()
}
\concept{auxiliary}
