% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-user.R
\name{rm_data_user}
\alias{rm_data_user}
\title{Extract and combine all user data}
\usage{
rm_data_user(login)
}
\arguments{
\item{login}{GitHub login of user}
}
\value{
A list of the following \code{data.frame} objects:
\enumerate{
\item \code{commit_cmt} with details of commits made on commits
\item \code{commits} with summaries of all repositories to which user made commits
\item \code{followers} A list of followers of specified user
\item \code{following} A list of other people who nominated user is following
\item \code{general} with some general information about specified user
\item \code{issue_cmts} with information on all issue comments made by user
\item \code{issues} with information on all issues opened by user
}
}
\description{
Extract and combine all user data
}
