% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githist-fn.R
\name{githist}
\alias{githist}
\title{Apply \pkg{pkgstats} across the git history of a package}
\usage{
githist(path, n = NULL, step_size = 1L)
}
\arguments{
\item{path}{Path to local repository containing an R package.}

\item{n}{If given, only analyses the preceding 'n' commits in the git
history.}

\item{step_size}{Analyse package for each \code{step_size} commits. For example,
with \code{step_size = 2}, analyses will be run on every second commit, instead
of default running on every commit. The value of \code{n} then analyses the
first \code{n} entries taken at intervals of \code{step_size}.}
}
\value{
A list of three items:
\itemize{
\item desc_data Containing data from \code{DESCRIPTION} files, along with data on
numbers of functions.
\item loc Containing data on "lines-of-code" for all languages and
sub-directories within package.
\item stats Containing statistics on (mean, medium, and sum) of various
properties of each function in package.
}
}
\description{
Apply \pkg{pkgstats} across the git history of a package
}
