% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto-dashboard.R
\name{repometrics_dashboard}
\alias{repometrics_dashboard}
\title{Start quarto dashboard with results of main \link{repometrics_data}
function.}
\usage{
repometrics_dashboard(data, action = "preview")
}
\arguments{
\item{data}{Results of main \link{repometrics_data} function applied
to one package.}

\item{action}{One of "preview", to start and open a live preview of the
dashboard website, or "render" to render a static version without previewing
or opening.}
}
\value{
(Invisibly) Path to main "index.html" document of quarto site. Note
that the site must be served with \code{action = "preview"}, and will not work by
simply opening this "index.html" file.
}
\description{
Start quarto dashboard with results of main \link{repometrics_data}
function.
}
