% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-repo.R
\name{repometrics_data}
\alias{repometrics_data}
\title{Collate 'repometrics' data for a local R package.}
\usage{
repometrics_data(
  path,
  step_days = 1L,
  num_cores = -1L,
  ended_at = Sys.time(),
  nyears = 1
)
}
\arguments{
\item{path}{Path to local repository containing an R package.}

\item{step_days}{Analyse package at intervals of this number of days. The
last commit for each day is chosen. For example, \code{step_days = 7L} will
return weekly statistics. Values of zero or less will analyse all commits,
including potentially multiple daily commits.}

\item{num_cores}{Number of cores to use in multi-core processing. Has no
effect on Windows operating systems, on which calculations are always
single-core only. Negative values are subtracted from number of available
cores, determined as \code{parallel::detectCores()}, so default of \code{num_cores = -1L} uses \code{detectCores() - 1L}. Positive values use precisely that number,
restricted to maximum available cores, and a value of zero will use all
available cores.}

\item{ended_at}{Parameter used in some aspects of resultant data to limit
the end date of data collection. Defaults to \code{Sys.time()}.}

\item{nyears}{Parameter <= 1 determining fraction of a year over which data
up until \code{end_date} are collected.}
}
\value{
A list of three forms of data:
\enumerate{
\item "pkgstats" containing statistics on the historical development of
package code, derived from the \pkg{pkgstats} package;
\item "rm" containing data from GitHub on the repository, including data on
contributors, issues, pull requests, and people watching and starring the
repository.
\item "contributors" as a named list of data on every individual contributor
to the repository, whether by code contributions or GitHub issues or
discussions.
}
}
\description{
This function collates all data for a local R package or repository needed
to create a dashboard with the \link{repometrics_dashboard} function. It
combines data from both the \link{repometrics_data_repo} and
\link{repometrics_data_user} functions.
}
\seealso{
Other data: 
\code{\link{repo_pkgstats_history}()},
\code{\link{repometrics_data_repo}()},
\code{\link{repometrics_data_user}()}
}
\concept{data}
