% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-user.R
\name{repometrics_data_user}
\alias{repometrics_data_user}
\title{Extract and combine all user data}
\usage{
repometrics_data_user(
  login,
  ended_at = Sys.time(),
  nyears = 1,
  n_per_page = 100
)
}
\arguments{
\item{login}{GitHub login of user}

\item{ended_at}{Parameter used in some aspects of resultant data to limit
the end date of data collection. Defaults to \code{Sys.time()}.}

\item{nyears}{Parameter <= 1 determining fraction of a year over which data
up until \code{end_date} are collected.}

\item{n_per_page}{Number of items per page to pass to GitHub GraphQL API
requests. This should never need to be changed.}
}
\value{
A list of the following \code{data.frame} objects:
\enumerate{
\item \code{commit_cmt} with details of commits made on commits
\item \code{commits} with summaries of all repositories to which user made commits
\item \code{followers} A list of followers of specified user
\item \code{following} A list of other people who nominated user is following
\item \code{general} with some general information about specified user
\item \code{issue_cmts} with information on all issue comments made by user
\item \code{issues} with information on all issues opened by user
}
}
\description{
Extract and combine all user data
}
