% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm-collate-all.R
\name{rm_chaoss_metrics_list}
\alias{rm_chaoss_metrics_list}
\title{List all implemented CHAOSS metrics}
\usage{
rm_chaoss_metrics_list()
}
\value{
A \code{data.frame} with two columns:
\enumerate{
\item "fn_names", with the internal function names of all implemented CHAOSS
metrics.
\item "url", with the URL to the CHAOSS community web page describing that
metric.
}
}
\description{
This function returns a list of internal functions defined within the
'repometrics' package. These internal functions are not intended to be
called directly, rather this list is provided for information only, to
enable users to know which metrics are implemented.
}
\note{
Metrics have been adapted in this package, and so may not precisely
reflect the descriptions provided in the CHAOSS community web pages linked
to in the URLs from this function. Adaptations have in particular been
implemented to align metrics with their usage in aggregate "models".
}
\examples{
metrics <- rm_chaoss_metrics_list ()
}
\concept{auxiliary}
